Move resolver::Context into new resolver::context module
